feat: Add guide number copy functionality
Introduces the ability to copy the document number (guia) from the BoletoCard component. This includes updating the UI to display the guide number and the associated copy button, as well as modifying the `handleCopy` function to accommodate this new type. Translations for the new copy functionality have also been added for both Portuguese and English.